move component re-rendering constant
