+++ src/js/ui/NewDataExploration/AdvancedConstraintCheckbox.js
@@ -60,7 +60,7 @@ import React, {useState} from 'react';
 import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
-import {componentIsRendering} from "../TabSystem";
+import {componentIsRendering} from "../Sidebar";
 import Util from "../../library/apertureUtil"
 
 function updateLayerConstraints(activeLayerConstraints, index) {
