+++ src/js/ui/NewDataExploration/WorkspaceSearchbar.js
@@ -65,7 +65,7 @@ import CheckBoxIcon from "@material-ui/icons/CheckBox";
 import {makeStyles} from "@material-ui/core/styles";
 import EqualizerIcon from "@material-ui/icons/Equalizer";
 import IconButton from "@material-ui/core/IconButton";
-import {componentIsRendering} from "../TabSystem";
+import {componentIsRendering} from "../Sidebar";
 import {isGraphable} from "./Helpers";
 import InfoIcon from '@material-ui/icons/Info';
 import {CustomTooltip} from "../UtilityComponents";
