+++ src/js/ui/TabSystem.js
@@ -68,7 +68,6 @@ import DataUsageIcon from '@material-ui/icons/DataUsage';
 import EqualizerIcon from "@material-ui/icons/Equalizer";
 import CloseIcon from "@material-ui/icons/Close";
 
-export const componentIsRendering = false;
 
 const useStyles = makeStyles((theme) => ({
     root: {
